move TargetPlatform to extensions (resolve cycle)
